Add additional settings controls, EditorPack settings listening, RenderPack selection
More clearly:

EditorPacks now receive 'global-conf-set' events when any global configuration is changed. This should be changed to allow for listening for specific keys. This is being used to set line wrapping on or off via the app footer.

RenderPacks can now be changed on the fly via a footer dropdown menu. Last RenderPack is saved.