Automatically run and use return values of Extension conf attributes
Extension configuration attributes that are functions that do not begin with "on" are automatically run and used as the attribute name's value when rendered to the screen. This enables element attributes to be dynamically defined based upon some arbitrary function provided by the user.

ome-ext-Colorizer uses this functionality to disable sections based on if use_system_colors is checked or not.

Also adjusted CSS pertaining to the disabled state.