Add DataManager, used for aggregrating data files
This simply provides methods for accessing files from multiple data
directories as if they were a single one. By setting up DataManager
with the following:

    DataManager.addPath(app.getPath('userData'));
    DataManager.addPath(app.getAppPath());
    DataManager.addPath(__dirname);

And given the following files on the system:

  * ~/.config/open-markup-editor/render-packs/ome-rp-custom
  * <APP>/render-packs/ome-rp-default
  * <CWD>/render-packs/ome-rp-custom2

DataManager.getFiles('render-packs', cb) would provide all files, with
absolute paths, in the different path's "render-packs" as an array to
the cb.