Implement file watching/reload settings & UI
Files may now be configured to not be watched at all, to automatically
reload a file if it was saved and the on-disk file changed, and/or to
prompt the user for a reload if the on-disk file has been changed.

This also includes a semi-iffy Core Settings extension pack that is used
to configure the above options.