Add additional git release parsing; add semver
Still more work to be done but we now support Gitea and GitHub. Others
should be possible as long as they provide a public releases API.

We now need to update the UI when there is a newer version within the
same major version(?) available and provide an update button. There
should probably be a step during installation to ensure that the OME
version is within an appropriate range, likely via a package.json "ome"
semver target.