Implement pack updating; Rework pack installation
Pack updating now works with github and gitea-like repositories! This
presumes that packs are released using semver tags in github/gitea
repositories.

Pack installation now unarchives the packs in a temporary directory
before attempting to install. This allows us, when updating, to
unarchive the new pack, move the old one to the temp, then move the new
in and install. If a failure happens during any of this, the old one can
simply be moved back in place.

Closes #20! :)