Implement pack type check; Adjust loading logic
During population, pack installation, and pack updating, we now attempt
to verify that a pack matches the appropriate type of the manager for
it. This prevents users from installing render packs as editor packs and
otherwise.

To have a pack verify as a type, a pack must either:

Have its directory name match a syntax such as '*-<PACKTYPE>-*',
where `<PACKTYPE>` must be any of the following:

  * **ext** - Extension
  * **rp** - Render Pack
  * **ep** - Editor Pack
  * **mp** - Markup Pack

Or have the following root entry in their package.json:

  * ome
    * packType: `<PACKTYPE>`

Where `<PACKTYPE>` is any of the following:

  * extension
  * editor
  * render
  * markup

Packs that do not match this will not be installed, and in the case of
an already installed pack, will be deleted from the packs directory.

---

Beyond this, async/await has been implemented in the initial pack
loading stages to make things less obtuse. However, there is some cruft
as a result of this kludge.